\name{analyzeMedianIncome}
\alias{analyzeMedianIncome}
\title{Analyse median Income statistically}
\usage{
  analyzeMedianIncome(demean = FALSE, division = FALSE,
    saveto = "~/Dropbox/mobility/output/data/census")
}
\arguments{
  \item{saveto}{}
}
\description{
  Analyse median Income statistically
}

