\name{backOutLevels}
\alias{backOutLevels}
\title{Back out price levels from MARSS simulation}
\usage{
  backOutLevels(l, n, N = 1, x0 = 1, saveto = NULL)
}
\arguments{
  \item{l}{the output of \code{\link{dynPrices}}}

  \item{n}{number of periods to simulate}

  \item{N}{number of different simulations}
}
\description{
  Back out price levels from MARSS simulation
}

