\name{buildDivDeviations}
\alias{buildDivDeviations}
\title{Build Deviations Models}
\usage{
  buildDivDeviations(n = 20, N = 2)
}
\description{
  Build Deviations Models
}
\seealso{
  Other ExpectationsModel:
  \code{\link{buildAggregatePrices}},
  \code{\link{estimateDivDeviations}},
  \code{\link{makeAggPricesFromData}},
  \code{\link{makeDivDifferences}},
  \code{\link{plotDivDeviations}},
  \code{\link{simDivDeviations}}
}

