\name{CreateBBLData}
\alias{CreateBBLData}
\title{Create the BBL data set}
\usage{
  CreateBBLData(data, RFmodels, BBLpars)
}
\arguments{
  \item{data}{full SIPP dataset}

  \item{RFmodels}{list with reduced form models}

  \item{BBLpars}{list with simulation parameters}
}
\description{
  forward simulate lifecycle profiles based on policy
  functions and initial conditions
}

