\name{DTSetChoice}
\alias{DTSetChoice}
\title{Set discrete choice on mover's histories}
\usage{
  DTSetChoice(DT)
}
\description{
  Set discrete choice on mover's histories
}
\seealso{
  Other LogitModel: \code{\link{buildLogit}},
  \code{\link{mergeHomeValues}},
  \code{\link{mergePredIncomeMovingHist}},
  \code{\link{printLogitModel}},
  \code{\link{simMovesFromLogit}}
}

