\name{estimateDivDeviations}
\alias{estimateDivDeviations}
\title{Estimate deviations models}
\usage{
  estimateDivDeviations(dat)
}
\description{
  Estimate deviations models
}
\seealso{
  Other ExpectationsModel:
  \code{\link{buildAggregatePrices}},
  \code{\link{buildDivDeviations}},
  \code{\link{makeDivDifferences}},
  \code{\link{plotDivDeviations}},
  \code{\link{simDivDeviations}}
}

