\name{getHomeValues}
\alias{getHomeValues}
\title{get home values and adjust by inflation}
\usage{
  getHomeValues(freq = "yearly")
}
\description{
  get home values and adjust by inflation
}
\seealso{
  Other IncomePrediction: \code{\link{RE.HHincome}},
  \code{\link{getIncomeM}}, \code{\link{getREintercept}},
  \code{\link{makePrediction1}},
  \code{\link{makePrediction2}}
}

