\name{getIncomeM}
\alias{getIncomeM}
\title{extract Xvars from data.table}
\usage{
  getIncomeM(with.FE, tmps)
}
\description{
  extract Xvars from data.table
}
\seealso{
  Other IncomePrediction: \code{\link{RE.HHincome}},
  \code{\link{getHomeValues}},
  \code{\link{getREintercept}},
  \code{\link{makePrediction1}},
  \code{\link{makePrediction2}}
}

