\name{getREintercept}
\alias{getREintercept}
\title{get RE intercept and merge into current state}
\usage{
  getREintercept(tmps, RE.coefs)
}
\description{
  get RE intercept and merge into current state
}
\seealso{
  Other IncomePrediction: \code{\link{RE.HHincome}},
  \code{\link{getHomeValues}}, \code{\link{getIncomeM}},
  \code{\link{makePrediction1}},
  \code{\link{makePrediction2}}
}

