\name{makeAggPricesFromData}
\alias{makeAggPricesFromData}
\title{make aggregate prices from real data}
\usage{
  makeAggPricesFromData(N)
}
\arguments{
  \item{N}{number of simulations}
}
\description{
  make aggregate prices from real data
}
\seealso{
  Other ExpectationsModel:
  \code{\link{buildAggregatePrices}},
  \code{\link{buildDivDeviations}},
  \code{\link{estimateDivDeviations}},
  \code{\link{makeDivDifferences}},
  \code{\link{plotDivDeviations}},
  \code{\link{simDivDeviations}}
}

