\name{makeHPIDivDifferences}
\alias{makeHPIDivDifferences}
\title{create dataset with HomeValue difference to national index}
\usage{
  makeHPIDivDifferences(d)
}
\description{
  create dataset with HomeValue difference to national
  index
}
\examples{
data(HomeValues,package="EconData")
makeHPIDivDifferences(dat)
}

