\name{makePrediction2}
\alias{makePrediction2}
\title{income prediction level 2}
\usage{
  makePrediction2(s, RE.coefs, m, with.FE, tmps,
    State_dist)
}
\description{
  predict income for all j given s
}
\seealso{
  Other IncomePrediction: \code{\link{RE.HHincome}},
  \code{\link{getHomeValues}}, \code{\link{getIncomeM}},
  \code{\link{getREintercept}},
  \code{\link{makePrediction1}}
}

