\name{mergeHomeValues}
\alias{mergeHomeValues}
\title{Merge Home Values with logit data}
\usage{
  mergeHomeValues(d)
}
\description{
  Merge Home Values with logit data
}
\seealso{
  Other LogitModel: \code{\link{DTSetChoice}},
  \code{\link{buildLogit}},
  \code{\link{mergePredIncomeMovingHist}},
  \code{\link{printLogitModel}},
  \code{\link{simMovesFromLogit}}
}

