\name{mergePredIncomeMovingHist}
\alias{mergePredIncomeMovingHist}
\title{Merge Moving Histories with Logit data}
\usage{
  mergePredIncomeMovingHist(l, mvt)
}
\description{
  Merge Moving Histories with Logit data
}
\seealso{
  Other LogitModel: \code{\link{DTSetChoice}},
  \code{\link{buildLogit}}, \code{\link{mergeHomeValues}},
  \code{\link{printLogitModel}},
  \code{\link{simMovesFromLogit}}
}

