\name{plot.CCPmoving}
\alias{plot.CCPmoving}
\title{Plot probability of moving}
\usage{
  plot.CCPmoving(m, b,
    path = "~/Dropbox/mobility/output/model/solution")
}
\arguments{
  \item{m}{model list}

  \item{b}{blitz solution list}

  \item{at}{at which slice to
  plotpath="~/git/migration/output/mgraphs/"}

  \item{path="~/git/migration/output/mgraphs/"}{}
}
\description{
  Plot the conditional choice probabilities of moving from
  locatino j to k. this is the function rho.
}

