\name{plot.TenureChoice}
\alias{plot.TenureChoice}
\title{plot savings and consumption functions}
\usage{
  plot.TenureChoice(m, b,
    path = "~/git/migration/output/mgraphs")
}
\arguments{
  \item{m}{list with model description}

  \item{b}{list of results from \code{\link{dev8}} Plot
  Tenure Choice Function}

  \item{m}{list with model description}

  \item{b}{list of results from \code{\link{dev8}}}
}
\description{
  plot policy functions at two sets of indices for owner
  and renter state
}

