\name{plot.vfuns}
\alias{plot.vfuns}
\title{plot value functions}
\usage{
  plot.vfuns(m, b, path = "~/git/migration/output/mgraphs")
}
\arguments{
  \item{m}{list with model description}

  \item{b}{list of results from \code{\link{dev8}}}
}
\description{
  plot value functions at two sets of indices for owner and
  renter state. Plots at 3 locations.
}

