\name{savingsModel}
\alias{savingsModel}
\title{Savings Policy Function Model}
\usage{
  savingsModel(d,
    saveto = "~/Dropbox/mobility/output/model/BBL/savings.RData")
}
\description{
  estimates the reduced form for liquid savings form data
}
\examples{
load("~/Dropbox/mobility/SIPP/Sipp4mn.RData")
}

