\name{simDivDeviations}
\alias{simDivDeviations}
\title{Simulate data from Deviation Models}
\usage{
  simDivDeviations(d, n = 20, N = 1, burn)
}
\description{
  Simulate data from Deviation Models
}
\seealso{
  Other ExpectationsModel:
  \code{\link{buildAggregatePrices}},
  \code{\link{buildDivDeviations}},
  \code{\link{estimateDivDeviations}},
  \code{\link{makeAggPricesFromData}},
  \code{\link{makeDivDifferences}},
  \code{\link{plotSimDivDeviations}}
}

