\name{makeDivDifferences}
\alias{makeDivDifferences}
\title{create dataset with devations from US average for income and prices}
\usage{
  makeDivDifferences()
}
\description{
  create dataset with devations from US average for income
  and prices
}
\examples{
data(US_medinc,package="EconData")
makeHPIDivDifferences()
}
\seealso{
  Other ExpectationsModel:
  \code{\link{buildAggregatePrices}},
  \code{\link{buildDivDeviations}},
  \code{\link{estimateDivDeviations}},
  \code{\link{plotDivDeviations}},
  \code{\link{simDivDeviations}}
}

