\name{plotDivDeviations}
\alias{plotDivDeviations}
\title{make plots of deviations models}
\usage{
  plotDivDeviations(n, N, path = NULL)
}
\description{
  make plots of deviations models
}
\seealso{
  Other ExpectationsModel:
  \code{\link{buildAggregatePrices}},
  \code{\link{buildDivDeviations}},
  \code{\link{estimateDivDeviations}},
  \code{\link{makeDivDifferences}},
  \code{\link{simDivDeviations}}
}

